fix #3254
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@25032 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netedit/GNEInspectorFrame.cpp
@@ -1006,7 +1006,7 @@ GNEInspectorFrame::AttributeInput::stripWhitespaceAfterComma(const std::string &
      * string = std::regex_replace(string, std::regex("\\$name"), "Somename");
      */
     std::string result;
-    for(int i = 0; i < (stringValue.size() - 1); i++) {
+    for(int i = 0; i < ((int)stringValue.size() - 1); i++) {
         if((stringValue.at(i) == ',') && (stringValue.at(i+1) == ' ')) {
             result.push_back(stringValue.at(i));
             i++;
